Split JavaOutput from CodeGenerators & Lexers
